Created registration activity for customer or store owner to create a new account.
Current design is that when creating a store owner account the user will create a customer account and then on the next activity fill in the store information.

When the user clicks the create new account button, either the data will be used to register them and then log them in after being validated, or if the checkbox button is filled they will be sent to the next activity to fill in the store information, after which the same thing will happen as with customer info except the store data will also be uploaded to the database.