Implemented the Store constructor and overrode equals and hashCode
The products of each store is stored in a HashSet as order does not matter. Moreover, it is not possible for a store to have duplicate products.

Overriden equals. 2 stores are equal if they have the same name and contain the same products. A store might have no products(NULL) but must have a name.

Overriden hashCode. If 2 store objects are equal then they return the same hashCode.