Added Resource object
It is not easy to retrieve resources from the xml files outside activity classes, so this class was created to help facilitate that. It is meant to be instantiated in the first activity and used throughout the lifecycle of the application.

This class should be used to retrieve resources anywhere that is not an activity.